+++ package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "@material-ui/core": "^4.11.0",
+    "@material-ui/icons": "^4.9.1",
     "apollo-datasource-rest": "^0.9.3",
     "apollo-server-micro": "2.11.0",
     "axios": "^0.19.2",
